backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
click==8.2.1
distro==1.9.0
fastapi==0.116.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==6.0.0
openai==1.93.2
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
xlsxwriter==3.2.5
SQLAlchemy>=1.4
passlib[bcrypt]>=1.7
python-jose>=3.3.0
python-dotenv>=1.1.1
email-validator>=1.3.0
passlib[bcrypt]>=1.7
cryptography>=3.4.7

